+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-13 03:12:48
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 7334e7007ab67580c6642f32f44c2ad45c4337f1
+
+## [ForwardWidgets-2kuai] - 2025-08-13 03:12:48
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 9b23dcdd3216c43ead46562a48cf18b7789bbc95
+
+## [ForwardWidgets-pack1r] - 2025-08-13 03:12:48
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 53d4782db3d0341991476b76d471ce9db1ec35e8
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-12 03:10:22
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main